+++ lib/biotcm/databases.rb
@@ -5,4 +5,5 @@
 module BioTCM::Databases
   autoload(:HGNC, "biotcm/databases/hgnc")
   autoload(:Cipher, "biotcm/databases/cipher")
+  autoload(:Medline, "biotcm/databases/medline")
 end